refactor: move PulseLive client to Integrations/PulseLive/
