+++ src/FplBot.Tests/Helpers/TestBuilder.cs
@@ -1,6 +1,6 @@
 using Fpl.Client;
 using Fpl.Client.Models;
-using Fpl.EventPublishers;
+using Fpl.PulseLive;
 using FplBot.Data;
 using FplBot.Data.Slack;
 using FplBot.Messaging.Contracts.Events.v1;
