+++ src/FplBot/Services/EventPublishers/FplWorkerServiceCollectionExtensions.cs
@@ -1,7 +1,6 @@
 using CronBackgroundServices;
-using Fpl.EventPublishers;
-using Fpl.EventPublishers.Abstractions;
 using Fpl.EventPublishers.Helpers;
+using Fpl.PulseLive;
 using Fpl.EventPublishers.RecurringActions;
 using Fpl.EventPublishers.States;
